update entity, debug crud account

+++ src/main/java/com/funtravelapp/payment/service/account/AccountService.java
@@ -1,5 +1,7 @@
 package com.funtravelapp.payment.service.account;
 
+import com.funtravelapp.payment.dto.account.CreateAccountRequest;
+import com.funtravelapp.payment.dto.account.CreateAccountValidator;
 import com.funtravelapp.payment.model.account.Account;
 import com.funtravelapp.payment.repository.account.AccountRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -12,9 +14,25 @@ import java.util.Optional;
 public class AccountService {
     @Autowired
     AccountRepository repository;
+    @Autowired
+    CreateAccountValidator createAccountValidator;
 
-    public Account create(Account account){
-        return repository.save(account);
+    public Account create(CreateAccountRequest request) throws Exception {
+        // Validation
+        boolean isValid = createAccountValidator.setRequest(request).validate();
+        if (!isValid){
+            throw new Exception("Request tidak lolos validasi");
+        }
+
+        Account acc = Account.builder()
+                .userId(request.getUserId())
+                .bank(request.getBank())
+                .number(request.getNumber())
+                .name(request.getName())
+                .type(request.getType())
+                .build();
+
+        return repository.save(acc);
     }
 
     public List<Account> getAll(){
@@ -27,11 +45,35 @@ public class AccountService {
         return opt.orElseThrow();
     }
 
-    public Account update(Account account){
-        return repository.save(account);
+    public Account update(int id, CreateAccountRequest request) throws Exception {
+        // Validation
+        boolean isValid = createAccountValidator.setRequest(request).validate();
+        if (!isValid){
+            throw new Exception("Request tidak lolos validasi");
+        }
+
+        boolean idExist = repository.existsById(id);
+        if (!idExist){
+            throw new Exception("Id tidak ditemukan");
+        }
+
+        Account acc = Account.builder()
+                .id(id)
+                .userId(request.getUserId())
+                .bank(request.getBank())
+                .number(request.getNumber())
+                .name(request.getName())
+                .type(request.getType())
+                .build();
+
+        return repository.save(acc);
     }
 
-    public void delete(int id){
+    public void delete(int id) throws Exception {
+        boolean idExist = repository.existsById(id);
+        if (!idExist){
+            throw new Exception("Id tidak ditemukan");
+        }
         repository.deleteById(id);
     }
